% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_L.R
\name{axis_bottom}
\alias{axis_bottom}
\title{Draw bottom axis only}
\usage{
axis_bottom(side, ..., line.col)
}
\arguments{
\item{side}{which side}

\item{line.col}{which line color}
}
\description{
Draw bottom axis only
}
