% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_L.R
\name{axis_C}
\alias{axis_C}
\title{Draw a C-shape axis framework}
\usage{
axis_C(side, ..., line.col)
}
\arguments{
\item{side}{which side}

\item{line.col}{which line color}
}
\description{
Draw a C-shape axis framework
}
