% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_L.R
\name{axis_L}
\alias{axis_L}
\title{Draw an L-shape axis framework}
\usage{
axis_L(side, ..., line.col)
}
\arguments{
\item{side}{which side}

\item{line.col}{which line color}
}
\description{
Draw an L-shape axis framework
}
