% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_L.R
\name{axis_left}
\alias{axis_left}
\title{Draw left axis only}
\usage{
axis_left(side, ..., line.col)
}
\arguments{
\item{side}{which side}

\item{line.col}{which line color}
}
\description{
Draw left axis only
}
