% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_ls.R
\name{axis_ls}
\alias{axis_ls}
\title{Control axis appearance in lattice plots}
\usage{
axis_ls(side, ..., line.col, type = "0011")
}
\arguments{
\item{side}{inherited from lattice (check if function works without this)}

\item{...}{parameter settings that are passed through}

\item{line.col}{set color of axis line}

\item{type}{character string specifying which lines should be drawn; see details below}
}
\description{
Control axis appearance in lattice plots
}
\details{
A four-number character string is used to specify which axes to draw, where a "1" means that the line will be drawn, "0" means it will be omitted. The sequence starts at the top (north -- east -- south -- west):
\itemize{
\item \code{"1111"} for all axes
\item \code{"0011"} for an L-shaped framework
\item \code{"1000"} for top axis only
\item \code{"0100"} for right axis only
\item \code{"0010"} for bottom axis only
\item \code{"0001"} for left axis only
\item \code{"1011"} for a C-shaped framework
\item \code{"0111"} for a U-shaped framework
}
}
