% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_L.R
\name{axis_U}
\alias{axis_U}
\title{Draw an U-shape axis framework}
\usage{
axis_U(side, ..., line.col)
}
\arguments{
\item{side}{which side}

\item{line.col}{which line color}
}
\description{
Draw an U-shape axis framework
}
