% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_ls}
\alias{col_ls}
\title{Custom color palette for plotting}
\format{
\subsection{\code{col_ls}}{

A vector with 8 colors:
}
}
\usage{
col_ls
}
\description{
This is a custom color palette.
}
\keyword{datasets}
