% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_to_dots.R
\name{dnorm_to_dots}
\alias{dnorm_to_dots}
\title{Create data points for a dot diagram representation of a normal density}
\usage{
dnorm_to_dots(n_dots = 100, mean = 0, sd = 1)
}
\arguments{
\item{n_dots}{number of dots}

\item{mean}{mean of normal distribution}

\item{sd}{standard deviation of normal distribution}
}
\value{
a numeric vector
}
\description{
Create data points for a dot diagram representation of a normal density
}
\examples{
dnorm_to_dots(
  n_dots = 30,
  mean = 0,
  sd = 1)
}
