% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fill_ls}
\alias{fill_ls}
\title{Custom fill colors for plotting}
\format{
\subsection{\code{col_ls}}{

A vector with 8 colors:
}
}
\usage{
fill_ls
}
\description{
This is a custom palette with fill colors.
}
\keyword{datasets}
