% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lattice_ls}
\alias{lattice_ls}
\title{Custom parameter settings for lattice}
\format{
\subsection{\code{lattice_ls}}{

A vector with 8 colors:
}
}
\usage{
lattice_ls
}
\description{
This is a list including custom parameter settings for plotting with the R package lattice.
}
\keyword{datasets}
