% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.dotdiagram.R
\name{panel.dotdiagram}
\alias{panel.dotdiagram}
\title{Draw a dot diagram in lattice}
\usage{
panel.dotdiagram(
  x,
  n_bins = 20,
  scale_y = 1,
  set_pch = 19,
  set_col = 1,
  set_cex = 1,
  set_alpha = 1,
  seq_min = min(x),
  seq_max = max(x),
  x_anchor = 0,
  y_anchor = 0,
  set_fontfamily = 1,
  set_lwd = 1,
  set_fontface = 1,
  vertical = F,
  leftwards = F,
  downwards = F
)
}
\arguments{
\item{x}{vector of values to be plotted}

\item{n_bins}{number of bins}

\item{scale_y}{controls the distance between stacked dots}

\item{set_pch}{plotting symbol}

\item{set_col}{color of plotting symbol}

\item{set_cex}{size of plottings symbol}

\item{set_alpha}{transparancy of plotting symbol}

\item{seq_min}{custom x-value to start binning}

\item{seq_max}{custom x-value to end binning}

\item{x_anchor}{for vertical dot diagram, base of the dots}

\item{y_anchor}{for dot diagram, base of the dots}

\item{set_fontfamily}{fontfamily of plotting symbol}

\item{set_lwd}{linewidth of plotting symbol}

\item{set_fontface}{fontface of plotting symbol}

\item{vertical}{whether the dotdiagram should be drawn vertically}

\item{leftwards}{whether the vertical dot diagram should be drawn leftwards}

\item{downwards}{whether the dot diagram should be drawn downwards}
}
\description{
Draw a dot diagram in lattice
}
