% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classic_ls.R
\name{theme_classic_ls}
\alias{theme_classic_ls}
\title{Modified ggplot2 "theme_classic()"}
\usage{
theme_classic_ls(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}

\item{base_line_size}{base line size}

\item{base_rect_size}{base rect size}
}
\description{
Modified ggplot2 "theme_classic()"
}
