% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dotplot.R
\name{theme_dotplot}
\alias{theme_dotplot}
\title{ggplot theme for dot plots}
\usage{
theme_dotplot(vertical = FALSE)
}
\arguments{
\item{vertical}{logical; whether the orientation of the dotplot is vertical}
}
\description{
ggplot theme for dot plots
}
